feature/search-filter-#129

+++ lib/domain/entities/product_entity.dart
@@ -2,8 +2,8 @@ class ProductEntity {
   final String id;
   final String name;
   final String description;
-  final int rating;
-  final int price;
+  final double rating;
+  final double price;
   final bool isAvailable;
   final String? photo;
   final List<String>? tags;
